data-plane/requirements.txt
flask
flask-cors
requests
pycryptodome
